Bugfix
I found the reason for the bug described in issue #1: the function to determine if a triangle is degenerate wasn't implemented properly. It checked only for triangles with two identical points, but it didn't check for triangles with collinear points. Collinear points result in faulty values for the angles. The proposed changes fix the problem.