More meaningful color choice
The choice of color for triangles is optimized in the change proposition:
The range of triangle errors is scanned prior to export. This allows to adjust the colors to represent the full spectrum of the values. Logarithm of values is used to enable a wider range of values to be represented. Faces with Illegal values (NaN/Infinite) are displayed in black/white to mark them as faulty.